Split install script into categories
With each category having its own install script, this will hopefully
make it easier to install individual pieces instead of the entire
system. For example, while testing a fresh Arch install on a new
machine, we can first install the minimal set of packages needed to
ensure that the machine has everything it needs instead of waiting for
all the programs to finish installing and then realizing that something
was missing in the configuration.