Update list of nvidia packages to install
This commit removes the nvidia-lts package to make sure that the nvidia
package is picked up since I'm using the linux kernel and not the
linux-lts kernel (the lts kernel with nvidia-lts was crashing pretty
often). This commit also installs the bbswitch package because
/usr/bin/prime-switch threw an error complaining that it didn't have
bbswitch. Hopefully this will fix all the crashing that I've been
experiencing.