Add awesome keybinding to (un)maximize client
Firefox started up in maximized mode and then I wasn't able to
unmaximize it. So I added this keybinding in for now. It seems like
Firefox preserves its own window settings on exit. I added the
keybinding in so that I could unmaximize it and so now it is not
maximized when I start it. Considering that things are working like I
want them to, I might remove this keybinding at some point, but it might
be useful to keep around just in case.