Update Arch install script
The Arch installer was running one pacman command per package that
needed to be installed. I tried this new method with my Ubuntu installer
and it seemed to work very well. Essentially, I can now maintain just a
list of packages in a file (as opposed to pacman install commands). I
then use a helper script to create the install command from the list of
packages which allows for one install command to be run per list of
packages which will hopefully speed things up and help catch install
errors a little easier.